[!!!][TASK] Refactor xml and bibtext export
Due to the new requirements of the ResponseInterface for controller actions
it is not possible "echo & exit" the response anymore.

Solution is to switch to page types for the downloads.

If sb modfied the templates, following things have to be done:

* add the typenums to partial "List/Export.html"
* change the file suffix from ".html" to ".xml" for the download templates

Additional (potential) break: Only the publications first CE on a page can
be exported.

Related #39

Signed-off-by: Marcus Schwemer <marcus.schwemer@in2code.de>