[FEATURE] Introduce option for author highlighting when searched
Now it is also possible to highlight the authors, which are used in the
search field in the filter. The option is set via flexforms in the tab
"Filter".
Due to the handling of flexforms was not possible to move the highlight
settings to the tab "Template", without breaking backwards compatibility.

Signed-off-by: Marcus Schwemer <marcus.schwemer@in2code.de>