[!!!][TASK] Reduce searched fields for searchterm filter
This many LIKE queries chained together for EACH
word in a searchterm query always caused a site
timeout either on webserver or database level.

An exact set of "recommended" fields to search
through for a generic term search isn't clear,
but I at least went ahead and removed the most
unlikely ones (identifiers should be their own
field).

In general, if someone wants a search this
complex, they should maybe use a search server
instead.

This is marked as "breaking" because it changes
the expected behaviour of the search compared to
previous versions.

Related: https://projekte.in2code.de/issues/71078